\name{chapterDescriptions}
\alias{chapterDescriptions}
\title{Chapter descriptions}
\usage{chapterDescriptions()}
\value{A data frame with chapter labels and short descriptions in chronological order.}
\description{Provides the chapter descriptions from the source contents. Missing descriptions are represented by \code{NA}.}
